work(task_engine): Change an assert to an if statement
I had a comment in _step_dfs that curr_step_name should always
name a valid step, which is why I made it an assert.  But that
turns out not to be true.  Authors can enter an invalid step as a
dependency, and that would violate the assumption.  So it needed
to change to a plain if statement, and raise a StepNotFound
exception.  Also add missing documentation to a docstring for a
function parameter.